chore: Added type checking via jsdoc and fixed types.
