+++ packages/gasket-plugin-express/package.json
@@ -52,10 +52,12 @@
     "eslint": "^8.56.0",
     "eslint-config-godaddy": "^7.0.2",
     "eslint-plugin-jest": "^27.6.3",
+    "eslint-plugin-jsdoc": "^48.2.0",
     "eslint-plugin-json": "^3.1.0",
     "eslint-plugin-unicorn": "^44.0.0",
     "express": "^4.18.2",
-    "jest": "^29.7.0"
+    "jest": "^29.7.0",
+    "typescript": "^5.3.3"
   },
   "peerDependencies": {
     "express": "^4.16.3"
@@ -63,13 +65,25 @@
   "eslintConfig": {
     "extends": [
       "godaddy",
-      "plugin:jest/recommended"
+      "plugin:jest/recommended",
+      "plugin:jsdoc/recommended-typescript-flavor"
     ],
     "plugins": [
-      "unicorn"
+      "unicorn",
+      "jsdoc"
     ],
     "rules": {
-      "unicorn/filename-case": "error"
+      "unicorn/filename-case": "error",
+      "valid-jsdoc": "off",
+      "spaced-comment": [
+        "error",
+        "always",
+        {
+          "markers": [
+            "/"
+          ]
+        }
+      ]
     }
   },
   "eslintIgnore": [
